requirements.txt
numpy==2.0.2
scikit-learn==1.6.1
pillow==11.2.1
pandas==2.2.3
seaborn==0.13.2
pyyaml==6.0.2
scipy==1.13.1
torch==2.7.0
torchvision==0.22.0
torchaudio==2.7.0
